refactor : HttpSession 폴더 이동
