+++ src/main/java/webserver/session/Attributes.java
@@ -1,4 +1,4 @@
-package http.common;
+package webserver.session;
 
 import java.util.HashMap;
 import java.util.Map;
