+++ src/test/java/webserver/session/HttpSessionTest.java
@@ -1,4 +1,4 @@
-package http.common;
+package webserver.session;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
